RelativeAgeEstimationAnonymous/Relative Age Estimation/E2ePipeline3/requirement.txt
absl-py==2.0.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blivet==3.6.0
Brlapi==0.8.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
cockpit @ file:///builddir/build/BUILD/cockpit-300.4/tmp/wheel/cockpit-300.4-py3-none-any.whl
comm==0.2.0
conda==4.14.0
conda-package-handling==1.7.3
condor-pytorch==1.1.0
contourpy==1.2.0
cryptography==36.0.1
cupshelpers==1.0
cycler==0.12.1
cytoolz==0.11.2
dasbus==1.4
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
distro==1.5.0
dnspython==2.3.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
file-magic==0.4.0
filelock==3.13.1
fonttools==4.47.0
fqdn==1.5.1
frozendict==1.2
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gpg==1.15.1
grpcio==1.60.0
gssapi==1.6.9
h5py==3.10.0
idna==2.10
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniparse==0.4
iotop==0.6
ipaclient==4.10.2
ipalib==4.10.2
ipaplatform==4.10.2
ipapython==4.10.2
ipykernel==6.28.0
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jwcrypto==0.8
kiwisolver==1.4.5
langtable==0.0.54
libcomps==0.1.18
lxml==4.6.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
netaddr==0.8.0
netifaces==0.10.6
networkx==3.2.1
nftables==0.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
perf==0.1
pexpect==4.8.0
pid==2.2.3
Pillow==10.1.0
platformdirs==4.1.0
ply==3.11
productmd==1.31
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.8.0
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.4
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycosat==0.6.3
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
Pygments==2.17.2
PyGObject==3.40.1
pyinotify==0.9.6
pykickstart==3.32.10
pyOpenSSL==21.0.0
pyparsing==3.1.1
pyparted==3.12.0
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-ldap==3.4.3
python-linux-procfs==0.7.0
python-meh==0.50
python-yubico==1.3.3
pytz==2021.1
pyudev==0.22.0
pyusb==1.0.2
PyYAML==5.4.1
pyzmq==25.1.2
qrcode==6.1
referencing==0.32.0
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rpm==4.16.1.3
rsa==4.9
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
scikit-learn==1.3.2
scipy==1.11.4
selinux==3.5
Send2Trash==1.8.2
sepolicy==3.5
setools==4.4.3
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.32/src
simpleline==1.8.3
six==1.15.0
sniffio==1.3.0
sos==4.6.1
soupsieve==2.5
SSSDConfig==2.9.1
stack-data==0.6.3
subscription-manager==1.29.38
sympy==1.12
systemd-python==234
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.11.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.65.0
traitlets==5.14.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==1.26.5
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
timm==1.0.15